Web/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pymongo==4.7.2
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.7.1
python-multipart==0.0.9
python-dotenv==1.0.0
Pillow==10.3.0
requests==2.31.0
aiosmtplib==3.0.1
email-validator==2.1.0
gradio_client>=2.0.0
